split refactor complete
